search: put predicate substitution logic in own package
