+++ internal/search/predicate/substitute.go
@@ -1,0 +1,151 @@
+package predicate
+
+import (
+	"github.com/grafana/regexp"
+
+	"github.com/sourcegraph/sourcegraph/internal/search/query"
+	"github.com/sourcegraph/sourcegraph/internal/search/result"
+	"github.com/sourcegraph/sourcegraph/lib/errors"
+)
+
+var ErrNoResults = errors.New("no results returned for predicate")
+
+// Substitute replaces all the predicates in a query with their expanded form. The predicates
+// are expanded using the doExpand function.
+func Substitute(q query.Basic, evaluate func(query.Predicate) (result.Matches, error)) (query.Plan, error) {
+	var topErr error
+	success := false
+	newQ := query.MapParameter(q.ToParseTree(), func(field, value string, neg bool, ann query.Annotation) query.Node {
+		orig := query.Parameter{
+			Field:      field,
+			Value:      value,
+			Negated:    neg,
+			Annotation: ann,
+		}
+
+		if !ann.Labels.IsSet(query.IsPredicate) {
+			return orig
+		}
+
+		if topErr != nil {
+			return orig
+		}
+
+		name, params := query.ParseAsPredicate(value)
+		predicate := query.DefaultPredicateRegistry.Get(field, name)
+		predicate.ParseParams(params)
+		matches, err := evaluate(predicate)
+		if err != nil {
+			topErr = err
+			return nil
+		}
+
+		var nodes []query.Node
+		switch predicate.Field() {
+		case query.FieldRepo:
+			nodes, err = searchResultsToRepoNodes(matches)
+			if err != nil {
+				topErr = err
+				return nil
+			}
+		case query.FieldFile:
+			nodes, err = searchResultsToFileNodes(matches)
+			if err != nil {
+				topErr = err
+				return nil
+			}
+		default:
+			topErr = errors.Errorf("unsupported predicate result type %q", predicate.Field())
+			return nil
+		}
+
+		// If no results are returned, we need to return a sentinel error rather
+		// than an empty expansion because an empty expansion means "everything"
+		// rather than "nothing".
+		if len(nodes) == 0 {
+			topErr = ErrNoResults
+			return nil
+		}
+
+		// A predicate was successfully evaluated and has results.
+		success = true
+
+		// No need to return an operator for only one result
+		if len(nodes) == 1 {
+			return nodes[0]
+		}
+
+		return query.Operator{
+			Kind:     query.Or,
+			Operands: nodes,
+		}
+	})
+
+	if topErr != nil || !success {
+		return nil, topErr
+	}
+	plan, err := query.ToPlan(query.Dnf(newQ))
+	if err != nil {
+		return nil, err
+	}
+	return plan, nil
+}
+
+// searchResultsToRepoNodes converts a set of search results into repository nodes
+// such that they can be used to replace a repository predicate
+func searchResultsToRepoNodes(matches []result.Match) ([]query.Node, error) {
+	nodes := make([]query.Node, 0, len(matches))
+	for _, match := range matches {
+		repoMatch, ok := match.(*result.RepoMatch)
+		if !ok {
+			return nil, errors.Errorf("expected type %T, but got %T", &result.RepoMatch{}, match)
+		}
+
+		repoFieldValue := "^" + regexp.QuoteMeta(string(repoMatch.Name)) + "$"
+		if repoMatch.Rev != "" {
+			repoFieldValue += "@" + repoMatch.Rev
+		}
+
+		nodes = append(nodes, query.Parameter{
+			Field: query.FieldRepo,
+			Value: repoFieldValue,
+		})
+	}
+
+	return nodes, nil
+}
+
+// searchResultsToFileNodes converts a set of search results into repo/file nodes so that they
+// can replace a file predicate
+func searchResultsToFileNodes(matches []result.Match) ([]query.Node, error) {
+	nodes := make([]query.Node, 0, len(matches))
+	for _, match := range matches {
+		fileMatch, ok := match.(*result.FileMatch)
+		if !ok {
+			return nil, errors.Errorf("expected type %T, but got %T", &result.FileMatch{}, match)
+		}
+
+		repoFieldValue := "^" + regexp.QuoteMeta(string(fileMatch.Repo.Name)) + "$"
+		if fileMatch.InputRev != nil {
+			repoFieldValue += "@" + *fileMatch.InputRev
+		}
+
+		// We create AND nodes to match both the repo and the file at the same time so
+		// we don't get files of the same name from different repositories.
+		nodes = append(nodes, query.Operator{
+			Kind: query.And,
+			Operands: []query.Node{
+				query.Parameter{
+					Field: query.FieldRepo,
+					Value: repoFieldValue,
+				},
+				query.Parameter{
+					Field: query.FieldFile,
+					Value: "^" + regexp.QuoteMeta(fileMatch.Path) + "$",
+				},
+			},
+		})
+	}
+
+	return nodes, nil
+}
